Actual page to load an individual camera.
The previous page was to list all the available cameras.

I’m just blown away at how easy it was to get a grouped dropdown list of all the available rolls sorted by their film types.

I realized that it should be a dropdown instead of just a <ul> so that I could more easily make it a form that posts to update the database. Going super old school with this to begin with. No JS.

https://docs.djangoproject.com/en/2.0/ref/templates/builtins/#regroup

https://docs.djangoproject.com/en/2.0/ref/templates/builtins/#grouping-on-other-properties