I guess we need empty values for empty options.
I was seeing “Push/Pull” as the value for roll.push_pull in the template when it wasn’t set to anything when submitting the form even though it still showed up as `-` in the admin. No idea why, but this seems to fix it and it looks like it’s how the placeholder `- - -` option is generated in a default Django choices form field.

https://docs.djangoproject.com/en/2.0/ref/models/fields/#choices