Make Camera names unique per user.
https://stackoverflow.com/a/16654393/96257
https://docs.djangoproject.com/en/2.0/ref/models/options/#unique-together

https://stackoverflow.com/a/7490785/96257

Got the idea to add `instance` to the edit view from here:
https://stackoverflow.com/a/526656/96257

Before I added that, I couldn’t edit a camera without changing the name I guess because it was validating it as if it were a new one with the same name as an existing instance?

Works now in any case.

I’m not sure if I should really have a separate Camera create and edit form/view. It does the trick for now.