Always add these fields.
I wish I had remembered. That was kind of confusing to figure out what
to do, although it ended up being pretty easy when I let the migration
command itself walk me through it.

I did find this:
https://www.reddit.com/r/django/comments/6zwx7c/wont_migrate_a_datetimefieldauto_now_addtrue/

Reddit of all things. Bleh.

Even though it looks like that default is what the model for good, the
command I followed was:

> 1) Provide a one-off default now (will be set on all existing rows)

And then:

> Please enter the default value now, as valid Python
> You can accept the default 'timezone.now' by pressing 'Enter' or you
> can provide another value.