Remove validation for a Roll's code.
It was validating that a Roll had a unique code per year for everyone,
not just the current user, which made it impossible for me to adjust a
code on one of @ploafmaster's rolls. It should be ok to let the system
handle this as it is now and only allow for modifying codes directly
via the Django admin, where said admin will hopefully be very careful.

Adding to my tests file to hopefully make sure things continue to
function as they should.