Start making the homepage look like the prototype.
- Putting this into a new branch for the time being since I'm being ugly
about tacking all this new content above all the existing stuff so I can
make sure all the data still is what it should be.
- Cleaned up the way I'm getting the variables for the context in the
index view and I have more to do to make all that consistent and pretty
compared to other places where I'm getting lists of rolls or counts.
- Also adding to the `/ready` page to make the "Outstanding" number make
sense. Making it work like my Film spreadsheet. Everything that's not in
Storage or Archived is "Outstanding."
- Installed Cypress to automatically take screenshots (with some
success). I'm choosing to leave it out of the codebase for the time
being (hence the additions to `.gitignore`).
- Adding some "help" text styling for a definition of "Outstanding."