Allow editing of code/label
Also:

- Allow for setting push_pull when adding non-storage roll
- Fix roll detail <title>

The idea with being able to edit the code is that it's not strictly
necessary and may or may not work for someone's preferred workflow.
We'll keep auto-generating, but we don't need to keep it so precious.

I may need to figure out how to make it unique per user per year,
though: https://github.com/trey/cassettenest/issues/22