Loaded camera link goes to journal entry page
Another little shortcut that came to mind from actually using the app.

I eventually want to be able to finish the roll directly from the
journal entry page as well. As a first pass, I just put a link to
the camera page which does have the finish roll button.