Be able to remove a camera from project (again)
Fixes #63

Don't allow someone to remove a camera loaded with a project-associated
roll because that would leave things in a weird state. It would still
show up as being associated with the project until the roll was finished
and then it would disappear (presumably). The loaded list shows cameras
based on their loaded rolls. Unloaded cameras show based on what's
associated with the project. It's kind of confusing.