Go back to project after loading a camera
Otherwise (if loading without associating a project), go to the roll
detail page (rather than the camera detail page).

Also require the roll field on camera load form. I know we need to
rethink how to do this (see #57), but this is a good stopgap.

Also show the camera name in the success message rather than just
"Camera loaded..."

Fixes #79