Timezone support!
Issues #116 and #61

- I decided it was ok to leave the default timezone as
  `America/New_York` since the database always uses UTC.
- New custom Profile model for adding extra info about users.
- New middleware to set the timezone to user settings.
- Display the timezone in the footer because we can.
- Customize the user admin a bit.

Here are a bunch of things that helped me figure this out.

I have found so many things on this website that have helped me out.

https://simpleisbetterthancomplex.com/tutorial/2016/07/22/how-to-extend-django-user-model.html#onetoone

The official documentation on how to deal with timezones in general.

https://docs.djangoproject.com/en/2.2/topics/i18n/timezones/#selecting-the-current-time-zone

The official documentation on how to extend the User model and customize
the default User admin.

https://docs.djangoproject.com/en/2.2/topics/auth/customizing/#extending-the-existing-user-model

Some good, but kind of flawed, ideas here. This is what got me started
down this path.

https://stackoverflow.com/questions/49914420/change-time-zone-depending-on-the-user-in-django-project

Good, general info on how Django middleware works.

https://simpleisbetterthancomplex.com/tutorial/2016/07/18/how-to-create-a-custom-django-middleware.html

This is how I figured out that you could define all the timezones in
your settings file. Found this by searching for
`django "timezone = models.CharField"`
Just trying to figure out a sensible way to define a `timezone` field on
the Profile model.

https://github.com/django-erp/django-erp/blob/f9621296a51d9b4e4120472ce29b6c29fc45d62e/djangoerp/settings/base.py.tmpl#L61

https://github.com/django-erp/django-erp/blob/f9621296a51d9b4e4120472ce29b6c29fc45d62e/djangoerp/core/models.py#L59

I need a constant reminder about `blank` and `null`.

https://simpleisbetterthancomplex.com/tips/2016/07/25/django-tip-8-blank-or-null.html

Troubles with error checking when a user doesn’t already have a profile.

https://stackoverflow.com/a/26270785/96257

https://stackoverflow.com/a/27064374/96257

Show the current timezone in a template.

https://docs.djangoproject.com/en/2.2/topics/i18n/timezones/#get-current-timezone