Save the working bits of PythonAnywhere wsgi's file to actual wsgi.py
Issue #117

It would be cool to get PythonAnywhere to actually run the version in
this project, but this is good for now. I'm learning a lot about how
these things are wired up, but I don't have the bravery to mess with
changing things again on a Friday morning that I'm supposed to be on
vacation.