Fix error when a user has multiple payment methods
Issue #43

Assuming a single payment Source for a user was a mistake. We need to
get at the `default_source` for the customer instead.

Sure wish this kind of thing had any documentation!

I sure hope it’s not a mistake to just have the exception catch a
Customer not existing. I guess we’ll soon see. I wonder if dj-stripe
uses signals on user creation or anything. Doesn’t look like it.