Add ability to update credit card
Issue #43

Moved bits of the subscription form into partials to include in both the
subscription page and the settings page. The forms are almost identical
and the JS is totally the same. The main difference is the backend code
for updating the card is a lot more simple than creating a customer,
adding a payment source, and then subscribing them to a plan. We’re
basically just adding a new default “source.”

Although, we are also deleting any existing payment sources for the
user. Had to go digging in the API and the dj-stripe source code to
figure out how to do that (although I did find the `detach` method in
the documentation after the fact).

Remind me to try to help improve the documentation in that project when
I get this all figured out.

https://stripe.com/docs/api/sources/detach

https://dj-stripe.readthedocs.io/en/stable/_modules/djstripe/models/payment_methods.html#Source.detach

https://dj-stripe.readthedocs.io/en/stable/reference/models.html#djstripe.models.Source.detach

https://dj-stripe.readthedocs.io/en/stable/reference/models.html#source