Link to film detail page from roll detail page
Fixes #144

Previously, both film detail and roll detail were using the same code to
display the manufacturer and name of film. Now the roll detail links to
the film detail which will then have the optional external links.

Eventually, we’ll make a manufacturer detail page that lists all of its
films and we’ll link to it from the roll detail page.