This was a change I made on December 27…
…but never committed. It was an idea from Django for Professionals
(which I got for Christmas). This change was sitting on Rosie and I
just wanted to get it out in the world because it was an unfinished
thought and it’s been about 7 weeks since I’ve committed anything to
this repo. :(

The thought behind this is that the way apps import is kind of
how CSS cascades so you want custom code to load after default Django
code. Here’s a quote from Django for Professionals page 17:

> Django loads apps from top to bottom so generally speaking it’s a
> good practice to add new apps below built-in apps they might rely on
> such as `admin`, `auth`, and all the rest.