Test for changing a roll from loaded to storage
This functionality already existed, but there was no test for it.

Next will be creating a test for changing the status from loaded to
anything else, which should automatically set its ended_on date and
unload its camera. And then I’ll write the functionality to make the
test pass.

Issue #139