Try to filter out spam subscriptions
I like the idea of having a visually hidden field and checking that
it’s still empty before passing it on to the Emma form. And if it’s
been filled, just redirect to the success page so the bot wouldn’t
know that it failed.

Honeypot.

Josh Mock recommended using the requests library, and I already had it
installed I think because of Stripe.