Camera detail view/template for backs or cameras
Start to revamp all the camera views to work for cameras or backs.

Had a blast from the past with the whole related_name thing. I knew it
seemed familiar: I had used that technique on ComicBinder. It took a
while to figure out how to get at the associated backs from a template
looking at a camera’s data. I worried for a bit that I wasn’t using
multiple-word models correctly, but I guess it’s ok. I guess I will
decide that it’s ok for the table name to not have an underscore. And
with the related_name thing, how you refer to a back is the same with
both camera and roll. All happy, right?

https://docs.djangoproject.com/en/3.0/topics/db/queries/#backwards-related-objects

It’s a little weird that the latest migration file was generated by
an older version of Django than the ones before it. I’m working on
Kuiil for the first time in a while and I forgot to update my pip
requirements.

Issue #11