Fix unloading a camera back, more consolidation
This feels like I’m right in the middle of a big mess. Even simplifying
things as much as I did with the Models (no many-to-many), getting
the views and templates consolidated (or not) feels like spaghetti code.
This would probably be a good use case for class-based views.

Issue #11