Try to get rid of a Docker warning.
I’m tired of seeing this message over and over again:

`The RELOAD variable is not set. Defaulting to a blank string.`

Leaving this with nothing after the `=` seems to work.

As a reminder, the reason I have this RELOAD thing set in the first
place is that so the app will reload when I make changes locally.