Add projects export.
I’m not sure about just outputting Camera and Roll IDs for this. At a
bare minimum, I want to be able to import this data back into the
system, but I want it to be as human readable as possible. I’ll have
to figure out what’s both functional and readable. Multi-dimension CSVs
seem super fancy, but I’m ready to learn.