Importing Cameras from a CSV mostly works!
I need to figure out if it’s possible to override the created and
updated with what’s in the CSV so we have total metadata integrity.

I also need to figure out if I need to handle this differently in the
case of a CSV file being enormous.

Here are some links that were helpful in getting to this point.

https://docs.djangoproject.com/en/3.1/topics/http/file-uploads/

https://medium.com/@simathapa111/how-to-upload-a-csv-file-in-django-3a0d6295f624

That Medium article is very similar to this YouTube video:

Django 2: How To Upload A CSV File To Django:
https://www.youtube.com/watch?v=BppyfPye8eo

I dig the simplicity and assuredness of this Stack Overflow answer, but
it wasn’t all that helpful.

python - Django: CSV model import:
https://stackoverflow.com/a/14158824

Which led to this official documentation. I like how you can call the
columns in the examples here by their names rather than index numbers.
https://docs.python.org/3.8/library/csv.html

This video (by the venerable Vitor Freitas) didn’t end up being super
helpful because I’m not trying to keep the uploaded files, just process
them for the sake of data import. I don’t think I need to do anything
with stuff like MEDIA_ROOT settings.

Introduction - Django File Upload Tutorial - Part 1:
https://www.youtube.com/watch?v=Zx09vcYq1oc

This Stack Overflow answer was most of the gist of the solution, but it
deals with a file path, and we don’t have it with the in-memory file or
something? This is new and exciting stuff, y’all.

python - how to import csv data into django models:
https://stackoverflow.com/a/15961736