Add a hopefully more functional error page.
That previous commit caused a server error and the 500.html template
was also giving the error, so it wasn’t doing much good. This may still
yet be too much logic to have in a 500 template, but we’ll see I guess.

I hope to never see this thing, but now if you do, it doesn’t look *as*
broken as a blank, white page with “Internal Server Error” on it.