Josh showed me how to parse arrays in a CSV.
Using `|` as a quote separator was silly. And silly for me to not
figure that out on my own. Now that it’s `"`, I can get at the entire
list of Roll IDs or whatever. That `json` bit is nice too. I would not
have found that on my own.

https://docs.python.org/3/library/json.html

Thank you, Josh Mock!