Add more model test coverage.
That `test_creating_profile_for_legacy_user` one was a huge rabbit
hole! Here are some things that helped me:

https://ponytech.net/blog/save-a-model-without-triggering-a-signal-nor-the-save-method

https://docs.djangoproject.com/en/3.0/ref/models/querysets/#bulk-create

That’s the same technique I use to keep the `created_at` and
`updated_at` fields to use what’s in the CSV when when import, but I
didn’t know you could do it for creating objects too. Neato.

I think this is what tipped me off to the existence of `assertRaises`:
https://github.com/django/django/blob/c70cd2a926ffab47f6613e83e0c8828eb6c2c064/tests/one_to_one/tests.py#L193

I came so close to just excluding the whole…

`except Profile.DoesNotExist:`

…stuff from the test coverage.
https://coverage.readthedocs.io/en/coverage-5.3/excluding.html#excluding-code-from-coverage-py