Test successful settings form post.
python - How can I unit test django messages?:
https://stackoverflow.com/a/46865530/

https://adamj.eu/tech/2020/06/15/how-to-unit-test-a-django-form/

Fixing the error message display to test against is being tricky.

The thing that’s going wrong is that I’m displaying them as `messages`
and it’s not picking up the logic for `form.errors`. The settings page
has 3 different forms on it, so I need to figure out a way to hook into
the same form error logic in _messages.html with multiple forms.

Maybe it’s time to start displaying some form errors inline with the
field instead of just all at the top of the page?