Add a test for roll exports.
This is very basic, but it’ll do for now. I eventually want to test
the data integrity, but I suppose that’ll happen more when I test the
imports.

Also found out that there were a couple of unnecessary imports in the
views.py file. I wish the linter would catch that kind of thing.

Count how many records are in a CSV Python?:
https://stackoverflow.com/a/16108605

python - How to test file response in Django?:
https://stackoverflow.com/a/47460338