Revamp overall site navigation around status.
On the logbook page, only show the year filter if we’re not already
filtering by status.

I was pretty much doing this in other places, but it’s nice for the
reassurance that it’s a good way to go about it.
https://tekshinobi.com/setting-active-navbar-link-in-django-template/

How I handle this kind of nav with 11ty and Nunjucks (so much nicer).
https://github.com/trey/treypiepmeier/blob/main/src/_includes/components/main-nav.njk

I owe this person so much for all help I’ve gotten:
https://simpleisbetterthancomplex.com/tips/2016/07/20/django-tip-7-how-to-get-the-current-url-within-a-django-template.html

https://docs.djangoproject.com/en/3.1/ref/request-response/#django.http.HttpRequest.META

python - How to See if a String Contains Another String in Django Template:
https://stackoverflow.com/a/19628156