Add tests for bulk roll updates.
Removing an `else` case from the rolls_update view since I was totally
unable to trigger it in a test. Hopefully that’s ok. And now the
rolls_update view has 100% test coverage.

Also making a couple of other minor adjustments to the rolls_update
view and the ReadyForm.

How to convert a Django QuerySet to a list:
https://stackoverflow.com/a/4424482