Get adding a subscription to work again!
The main thing that was broken was fixed by putting this in settings.py:
STRIPE_API_VERSION = '2020-03-02'

Adding a card fails in 2.4.1:
https://github.com/dj-stripe/dj-stripe/issues/1293#issuecomment-739123493

Finally found this solution by searching for “stripe_object.py
attributeerror” and finding this page:
https://www.bountysource.com/issues/94969855-adding-a-card-fails-in-2-4-1

That was aggravating as hell and impossible for me to figure out on my
own. I’m seriously considering simplifying things and using Stripe
Checkout instead of having the forms live on my site.

---

The update to views.py came from here:
https://github.com/dj-stripe/dj-stripe/blob/44fae4e86ae43c421fc4b43fbfd6f9b2e4d6ac91/tests/apps/example/views.py#L75