Start writing backup verification scripts.
I haven’t tried these out yet, but I think I’m on the right track.

I still want to read a bit more about the best way to wipe a database
and replace it with a different version. This is a hard thing to find
information about on the internet apparently.

I’ve changed the file names for daily backups to remove the hours,
minutes, and seconds for the sake of an easier script that doesn’t need
to get a full list of the files on Spaces, parse out the file name
using regular expressions, and then use that to actually download the
file from Spaces. And, since these files should only be created once a
day, they don’t need to have more granular filenames. The code-push
ones can stay the way they are for now.