Fix script.
That was a fun way to totally lose my local database. Better the local
than the production. I was able to grab one from another machine anyway.
This happened just because I forgot the `$` on that `mkdir` command. I
guess `s3cmd` will create a folder if it needs to, but `pg_dump` won’t?

It seems like deleting the local.dump file was redundant anyway, so I’m
glad to leave that out.