Add optional trial subscription mode.
I felt compelled to give this a shot. I was surprised at how
straightforward it ended up being, but it did eat up a lot of a day.
I’m tickled with how it came out, though. Now all I have to do to
enable it is to change one or two values in the `.env` file. Well, and
then write all the logic to put an account into read-only mode and
require a subscription upon registration. But at least the trial piece
of the work is done now.

General Checkout stuff:
https://stripe.com/docs/billing/subscriptions/checkout

Trial specifics:
https://stripe.com/docs/billing/subscriptions/trials

How to set a trial period:
https://stripe.com/docs/api/checkout/sessions/create#create_checkout_session-subscription_data-trial_end

`trailing` status:
https://stripe.com/docs/billing/subscriptions/overview#subscription-statuses

When a subscription goes from `trailing` to `active`, our webhook
should update the user automatically.
https://stripe.com/docs/api/subscriptions/object?lang=python#subscription_object-status

Days between two dates?:
https://stackoverflow.com/a/8258465