Finally update the Profile model.
I decided to just start deleting all the Stripe stuff. It was going to
be a headache to try to translate between the different status names
and all that.

Also deleting all the trial stuff and will re-write it after I get the
base level subscription stuff working again.

I always need a reminder about when to use `null=True`.
https://simpleisbetterthancomplex.com/tips/2016/07/25/django-tip-8-blank-or-null.html

How to compare dates in Django templates:
https://stackoverflow.com/a/34959070