Add Radio form field option.
This has been done for a while in the pattern repository, but I
apparently never added it here. I tried using it for the
change-your-plan options, but I ended up wanting to customize it more
than would be reasonable using the Django “component,” so we’ll just
leave this here for now. I’m sure it’ll come in handy at some point.

https://github.com/trey/cn-storybook/commit/62d392201a542f97078417836d7c5831666c63c1

https://github.com/trey/cassettenest/commit/76cd0b10f92fb7be52b5ebb4223d235c451585ad