Update plan display on page load.
So if you select a plan you don’t currently have and then refresh the
page, it still shows what was there when you clicked it. This is kind
of a little bit of state management, no? So glad the only state that
matters here is on the server. People are absolutely banana factories
for making single-page apps. Turbo bananas.