Try out a GitHub Action.
I have no idea what I'm doing. This is a combination of the default
stuff that comes up when you go to create a workflow and the example 
usage from https://github.com/dc740/django-coveragepy-github-action

I found that Action by searching the “Marketplace” in the sidebar of
the create an action page.