Fancy Markdown list autocomplete.
I spent all day working on this, and I’ll bet someone has published a
way to do this and I just wasn’t able to find it. I know GitHub has a
way to do it, but it seems as if they haven’t published it yet.

This was a good learning experience. Is what I’ll tell myself.

It took me the longest time to get an inkling of how cursor position
works in a textarea. It would be nice if it were x and y coordinates,
but I guess maybe it's number of characters?

https://github.blog/changelog/2020-12-15-markdown-list-syntax-now-autocompleted/

https://www.geeksforgeeks.org/how-to-insert-text-into-the-textarea-at-the-current-cursor-position/

jQuery Set Cursor Position in Text Area:
https://stackoverflow.com/a/499158

https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement/setSelectionRange