Fix bug with saving a 35mm roll loaded in a back.
That view was a total mess. This is a lot better.

What was happening was that the form never showed the camera that was
loaded if it was a medium format camera with a 35mm roll of film
because it wouldn’t populate a different format camera. And then it
totally relied on what came back in the form to save the record.

So now we’re trusting the form to do its own thing other than setting
camera / back to what was already there if it doesn’t come through the
form. And we don’t let you change the camera for a loaded roll.