Update npm, pin Node version more specifically.
The package-lock changes happened just from my running `npm i` after
pulling down the changes from Dependabot. It was because I had updated
npm on my machine for the sake of working with Marc (from 6.14.7 to
7.13.0). Apparently the whole lockfileVersion change from 1 to 2
happened with version 7 of npm:
https://docs.npmjs.com/cli/v7/configuring-npm/package-lock-json#lockfileversion

So I decided to make a commit of being more specific with nvm and
explaining what happened rather than just have a commit that was only a
change to package-lock.

All this is for the sake of having linting for Sass and JavaScript!
Node is terrible and I do not like it.

Hopefully Dependabot will be able to keep things from changing again
noting that the lockfile is set to version 2 now.