Adjust models to start to depreciate duplicates.
The idea is to make these things optional and then only show them in
the UI if the value exists. Then we can phase out Film having that info
and rely instead on Stocks.

I was trying to make the auto migration a little prettier, but it
really didn’t want me to change its terrible string formatting
apparently.