Adding a stock actually works. Revamped stuff.
This is a mess of a commit that I started yesterday and finished today.

I’ll need to go ahead and create stocks for every film in the system.
Shouldn’t be too hard, I don’t think. And nobody’s added a new film in
quite a while, so it shouldn’t be too hard to create those and point
them in the right places.

- Renaming lots of things.
- Removing duplication of format names.
- Realizing that for the sake of adding films to inventory or logbook,
  there’s no backwards compatibility possible (that I can quickly
  figure out) for film.manufacturer. Gotta go all-in on
  stock.manufacturer.
- Prevent logbook add form from having access to EVERYBODY’s film, even
  ones marked personal. Whoops.
- Getting tests to pass again.
  I wish I could require film.stock, but I think that would screw up
  things royally until all the new stocks were added. I wonder if I can
  make that change after making sure every film has a stock.

Python string.join(list) on object array rather than string array:
https://stackoverflow.com/a/497773