Make sure type dropdown resets when needed.
Hitting the browser back/forward buttons from a page that had a single
type (or no types) wouldn’t reset to show the correct types on another
page.

I find this syntax a little puzzling, but I think it’s equivalent to
calling `changed=changed`, but since it’s a `const` the browser was
fussing at me.