Add a fallback for type being 'null'.
I’m not sure how that’s happening, but it shouldn’t ever make the
server throw up if it does.

It seems to happen every once in a while when I’m clicking the back and
forward buttons in the browser very quickly. Must be some kind of a
race condition, but I don’t know how this line isn’t catching it:

```javascript
const newType = (url.searchParams.get('type')) ? url.searchParams.get('type') : 'all';
```

If it’s `null`, that ternary should be false, right?