Reset type selection when changing manufacturer.
As an example: previously, if you selected c41 and then changed to
Ilford, c41 would still be selected and there would be nothing showing.
This was a bug caused by having everything reset every time for the
benefit of the back/forward button stuff. Now everything seems to work
as expected.

I really need to get some JS testing in place.