Fix back/forward buttons again.
We need to assume type (and therefore everything) changes on popstate
so that it will match the URL above all else.

What was happening this time was that if you, for example, went to
CineStill and then chose c41, hit back to all types, and then forward
to c41 again, it would show all types in the list but the select would
still have c41 selected. This seems to fix that.

I noticed this issue when playing around on my iPhone.

Committing this from the Fairfield Inn & Suites by Marriott
Williamstown, MA.