Remove some Sass stuff.
All of the Sass stuff? Hopefully that's all but the files themselves.

Also adding in some of the Tailwind generation stuff we'll be needing.

I wasn't sure if I should remove all the stuff under
`STATICFILES_FINDERS`, but I figured it would be ok to set it to the
defaults (and using a list instead of a tuple) since that's what in
the cn-tools project. I would have just fired up a new Django project
and seen what it includes, but I can't remember how to do that off the
top of my head and it was late when I was doing this.
https://docs.djangoproject.com/en/3.2/ref/settings/#staticfiles-finders