Structure Ajax calls a bit better.
Things are temporarily broken with this.

This establishes a new url/view for the sake of these Ajax requests.
This will make it easier to show the correct URL in the browser and
load the correct data into the correct section. This feels more complex
than it ought to be, but it will be awesome.

This is good stuff:
https://www.mattlayman.com/blog/2021/how-to-htmx-django/