Upgrade Django Database Backup.
I decided it's time to get this thing updated even if it's still "beta"
so I can go ahead and upgrade to Django 4. Wish me luck.

I think what I'll do is upgrade this, have it run its daily backups for
a day or so and do a local check to see if it's working alright. I can
always roll it back to a previous version if necessary I guess. That'd
be quite an ordeal if I'd upgraded Django too. It'll be alright.