Probably a mistake, but now my tests work so…
Here's what Adam Johnson says (Improve Your Django DX page 215):

> …we want to ignore any local .env file during tests, to ensure our
> test environment is repeatable and isolated. In the worst case, tests
> reading a .env file could configure a remote API integration, and the
> tests could make unintentional requests to the live API.

I guess we can keep that in mind but just keep rocking for now. There
were just a ton of little things that I had to do to get it to even
start testing and then a bunch of things failed because they were
missing Paddle IDs and stuff. And to get the database to even start, I
had to add this right after the regular database configuration in
settings.py:

if command == 'test':
    DATABASES['default']['NAME'] = 'cassette_nest'