Remove an extraneous halt event from this button.
This was not worth the trouble. Especially since this whole UI needs to
be redone anyway. I had to wrap this thing in a parent div because I
couldn't override the button's inline-flex with a `hidden` class on the
element itself. How does Tailwind not allow that kind of override?