Fix dark mode JS.
Fixes piepworks/cassettenest-roadmap#

It was breaking when using back and forward browser buttons + htmx.

And that, in turn was preventing the htmx indicator's opacity from going
back to 1.0.

Also making that field disappear from settings without JS as it was
before. That must have been a goof to have removed it.