Test registration
This is still very basic. Just getting a feel for how to use Playwright
with the test server (instead of a locally running server like the
examples all show) so that this is ready to be plugged into a GitHub
workflow.

https://stackoverflow.com/questions/2030053/how-to-generate-random-strings-in-python

Started out with this stuff:
https://playwright.dev/python/docs/codegen-intro

Another great example from this repo:
https://github.com/Chris-May/django_playwright_pytest_example/blob/eec89612e36ace05fb80874bf5f2f147a3f2a590/tests/test_ui.py#L12