Set up auth in Playwright, fix patterns page
This was a heck of a thing to do. More yak shaving than one should have
to do.

The patterns page had some dynamic stuff that needed to go away to not
trigger a difference in screenshots.

Fall back on dotenv if environment variables aren't set directly.
Basically only for local use since we're gonna want to create a user
every time tests run on GitHub

Also converting things from CommonJS to ES modules because apparently
that's the more standard way of doing things nowadays. Thank you,
VSCodium for pointing it out and me for doing a tiny bit of research to
try to understand it:
https://blog.logrocket.com/commonjs-vs-es-modules-node-js/

Link dump:

https://playwright.dev/docs/auth

How to have certain tests be un-authenticated:
https://github.com/microsoft/playwright/issues/11809#issuecomment-1146575789

https://playwright.dev/docs/test-parameterize#env-files
https://playwright.dev/docs/test-parameterize#passing-environment-variables

Good refresher on `createsuperuser --noinput`
https://stackoverflow.com/a/59467533

https://docs.djangoproject.com/en/stable/ref/django-admin/#createsuperuser