Re-break the "double bounce on htmx selects" :(
piepworks/cassettenest-roadmap#312

It was causing a bug where when you reset filters, the selects didn't
update (to show the "all" option on everything) because `hx-preserve`
always keeps that element from updating.

We'll have to figure out another way to do this. I intend to look at
some alternatives to htmx anyway.