Fix bug when pushing a roll
If it was a push and not a pull, it was just storing a positive integer
instead of the string "+whatever".

I can't believe that nobody has tried to push a roll of film since I
introduced that stepper component.

python - Checking whether a variable is an integer or not:
https://stackoverflow.com/a/3501408