Revamp film/roll picker fields
There was no reason to be using a custom template just for those.

This was a rabbit hole I went down trying to work on fixing the order of
cameras on the add-a-roll-to-logbook page:
piepworks/cassettenest-roadmap#194

I started working on a more standard grouped dropdown template and then
realized I had already done something more sophisticated based on this:
https://simpleisbetterthancomplex.com/tutorial/2019/01/02/how-to-implement-grouped-model-choice-field.html

It feels a little crude that I'm basically duplicating the logic in the
new fields.py just changing a bit to show counts in some places, but
this is a lot nicer to just be able to use form-field.py in all the
places. Optimizing for cleaner templates is a good thing, right?

Now onto incorporating this logic in all the camera picker form fields.

Also made the form labels more consistent and had to regenerate
screenshots as a result.

https://playwright.dev/docs/test-snapshots
`npx playwright test --update-snapshots`