Update the camera picker on project detail page
I have no idea why I had to update the Playwright screenshots again. I
tried deleting all the compiled things and saw that it needed to run
`build` to generate the `tailwind.css` file, so that's why that's been
added to the `start-django` script.

How can I capitalize the first letter of each word in a string?:
https://stackoverflow.com/a/1549644