Add grain from marketing site & a light version
I can already tell we're going to need to make some adjustments to the
sidebar if nothing else. Maybe the dark grain will look good there.

I'm seeing this commit and the `grain` branch as the start of backing
out of requiring subscriptions (or even offering them) and starting to
promote donations and keeping track of your own donations in the app.

piepworks/cassettenest-roadmap#326